i#2350 rseq: Refactor rseq code into its own file
In preparation for future work which will add additional
per-rseq-region data and code, moves the rseq code into its own file
core/unix/rseq_linux.c.  Adds separately-allocated data to each rseq
region, with an interface for querying it, to facilitate further data
in the future.

Issue: #2350
